test(audit): split recovery identity coverage
