+++ src/agents/main-session-recovery-state.test.ts
@@ -8,18 +8,6 @@ import { projectMainSessionRecoveryLifecycle } from "./main-session-recovery-lif
 import { transitionMainSessionRecovery } from "./main-session-recovery-state.js";
 
 const sessionKey = "agent:main:main";
-const executionIdentity = (runId: string) => ({
-  tokenVersion: 1 as const,
-  contextId: `context-${runId}`,
-  executionId: `execution-${runId}`,
-  runId,
-  createdAt: 1,
-});
-const enabledExecutionIdentity = (runId: string) => ({
-  state: "enabled" as const,
-  token: executionIdentity(runId),
-});
-
 function recoveryState(
   overrides: Partial<MainRestartRecoveryState> = {},
 ): MainRestartRecoveryState {
@@ -175,110 +163,6 @@ describe("main session recovery state", () => {
     ]);
   });
 
-  it("charges at reservation and refunds only the matching reservation", () => {
-    const entry = interruptedEntry();
-    const prepared = transitionMainSessionRecovery(entry, {
-      kind: "prepare_attempt",
-      attempt: 1,
-      lifecycleGeneration: "generation-1",
-      now: 200,
-      observation: { sessionId: "session-1", cycleId: "cycle-1", revision: 1 },
-      runId: "recovery-1",
-      executionIdentity: enabledExecutionIdentity("recovery-1"),
-    });
-    expect(prepared.kind).toBe("reserved");
-    if (prepared.kind !== "reserved") {
-      throw new Error("expected reservation");
-    }
-    expect(prepared.reservation).toMatchObject({
-      executionIdentityAdmission: {
-        kind: "capture",
-        token: executionIdentity("recovery-1"),
-      },
-    });
-
-    expect(
-      transitionMainSessionRecovery(entry, {
-        kind: "prepare_attempt",
-        attempt: 1,
-        lifecycleGeneration: "generation-1",
-        now: 201,
-        observation: { sessionId: "session-1", cycleId: "cycle-1", revision: 1 },
-        runId: "recovery-2",
-        executionIdentity: enabledExecutionIdentity("recovery-2"),
-      }),
-    ).toEqual({ kind: "rejected", reason: "stale_revision" });
-    expect(entry.mainRestartRecovery?.reservation).toMatchObject({
-      runId: "recovery-1",
-      attempt: 1,
-    });
-
-    const claim = claimForeground(entry);
-    expect(claim.kind).toBe("foreground_claimed");
-
-    expect(
-      transitionMainSessionRecovery(entry, {
-        kind: "cancel_reservation",
-        reservation: prepared.reservation,
-      }),
-    ).toEqual({ kind: "applied" });
-    expect(entry.mainRestartRecovery).toMatchObject({
-      chargedAttempts: 0,
-      foregroundClaims: {
-        lifecycleGeneration: "generation-1",
-        tokens: ["foreground-1"],
-      },
-    });
-    expect(entry.mainRestartRecovery?.reservation).toBeUndefined();
-    expect(entry.mainRestartRecovery?.executionIdentity).toBeUndefined();
-    expect(observe(entry, "generation-1")).toEqual({ status: "blocked" });
-  });
-
-  it("keeps disabled recovery identity out of durable state and reservations", () => {
-    const entry = interruptedEntry();
-
-    const prepared = transitionMainSessionRecovery(entry, {
-      kind: "prepare_attempt",
-      attempt: 1,
-      lifecycleGeneration: "generation-1",
-      now: 200,
-      observation: { sessionId: "session-1", cycleId: "cycle-1", revision: 1 },
-      runId: "recovery-1",
-      executionIdentity: { state: "disabled" },
-    });
-
-    expect(prepared).toMatchObject({ kind: "reserved" });
-    if (prepared.kind !== "reserved") {
-      throw new Error("expected reservation");
-    }
-    expect(prepared.reservation.executionIdentityAdmission).toBeUndefined();
-    expect(entry.mainRestartRecovery?.executionIdentity).toBeUndefined();
-  });
-
-  it("does not propagate a previously retained token while collection is disabled", () => {
-    const retained = executionIdentity("recovery-1");
-    const entry = interruptedEntry({
-      mainRestartRecovery: recoveryState({ executionIdentity: retained }),
-    });
-
-    const prepared = transitionMainSessionRecovery(entry, {
-      kind: "prepare_attempt",
-      attempt: 1,
-      lifecycleGeneration: "generation-1",
-      now: 200,
-      observation: { sessionId: "session-1", cycleId: "cycle-1", revision: 1 },
-      runId: "recovery-1",
-      executionIdentity: { state: "disabled" },
-    });
-
-    expect(prepared).toMatchObject({ kind: "reserved" });
-    if (prepared.kind !== "reserved") {
-      throw new Error("expected reservation");
-    }
-    expect(prepared.reservation.executionIdentityAdmission).toBeUndefined();
-    expect(entry.mainRestartRecovery?.executionIdentity).toEqual(retained);
-  });
-
   it("rejects foreground work after the automatic recovery budget is exhausted", () => {
     const entry = interruptedEntry({
       mainRestartRecovery: recoveryState({ chargedAttempts: 3 }),
@@ -411,63 +295,6 @@ describe("main session recovery state", () => {
       },
     });
     expect(entry.mainRestartRecovery?.reservation).toBeUndefined();
-  });
-
-  it("releases an ambiguous dispatch reservation without refunding its charge", () => {
-    const entry = interruptedEntry();
-    const prepared = transitionMainSessionRecovery(entry, {
-      kind: "prepare_attempt",
-      attempt: 1,
-      lifecycleGeneration: "generation-1",
-      now: 200,
-      observation: { sessionId: "session-1", cycleId: "cycle-1", revision: 1 },
-      runId: "recovery-1",
-      executionIdentity: enabledExecutionIdentity("recovery-1"),
-    });
-    if (prepared.kind !== "reserved") {
-      throw new Error("expected reservation");
-    }
-
-    expect(
-      transitionMainSessionRecovery(entry, {
-        kind: "abandon_reservation",
-        reservation: prepared.reservation,
-      }),
-    ).toEqual({ kind: "applied" });
-    expect(entry.mainRestartRecovery).toMatchObject({ chargedAttempts: 1 });
-    expect(entry.mainRestartRecovery?.reservation).toBeUndefined();
-    expect(observe(entry, "generation-1")).toMatchObject({
-      status: "recoverable",
-      nextAttempt: 2,
-    });
-    const retry = transitionMainSessionRecovery(entry, {
-      kind: "prepare_attempt",
-      attempt: 2,
-      lifecycleGeneration: "generation-1",
-      now: 300,
-      observation: {
-        sessionId: "session-1",
-        cycleId: "cycle-1",
-        revision: entry.mainRestartRecovery!.revision,
-      },
-      runId: "recovery-1",
-      executionIdentity: {
-        state: "enabled",
-        token: {
-          ...executionIdentity("replacement"),
-          runId: "recovery-1",
-        },
-      },
-    });
-    expect(retry).toMatchObject({
-      kind: "reserved",
-      reservation: {
-        executionIdentityAdmission: {
-          kind: "retry-reference",
-          token: executionIdentity("recovery-1"),
-        },
-      },
-    });
   });
 
   it("moves a reservation into the lifecycle fence during Gateway admission", () => {
@@ -707,7 +534,7 @@ describe("main session recovery state", () => {
         now: 500,
         observation: oldObservation,
         runId: "stale-run",
-        executionIdentity: enabledExecutionIdentity("stale-run"),
+        executionIdentity: { state: "disabled" },
       }),
     ).toEqual({ kind: "rejected", reason: "stale_cycle" });
   });
